[form-factors] Move BFvD2014 parametrization to separate files
