+++ eos/form-factors/baryonic_TEST.cc
@@ -27,59 +27,6 @@
 using namespace test;
 using namespace eos;
 
-class BFvD2014FormFactorsTest :
-    public TestCase
-{
-    public:
-        BFvD2014FormFactorsTest() :
-            TestCase("bfvd2014_form_factors_test")
-        {
-        }
-
-        virtual void run() const
-        {
-            static const double eps = 1e-3;
-
-            Parameters p = Parameters::Defaults();
-            std::shared_ptr<FormFactors<OneHalfPlusToOneHalfPlus>> ff = FormFactorFactory<OneHalfPlusToOneHalfPlus>::create("Lambda_b->Lambda::BFvD2014", p, Options{ });
-
-            p["Lambda_b->Lambda::f_0^V(0)@BFvD2014"]    =  0.33;
-            p["Lambda_b->Lambda::b_1_0^V@BFvD2014"]     = -1.75;
-            p["Lambda_b->Lambda::f_0^A(0)@BFvD2014"]    =  0.31;
-            p["Lambda_b->Lambda::b_1_0^A@BFvD2014"]     = -0.52;
-            p["Lambda_b->Lambda::f_perp^V(0)@BFvD2014"] =  0.34;
-            p["Lambda_b->Lambda::b_1_perp^V@BFvD2014"]  = -1.58;
-            p["Lambda_b->Lambda::f_perp^A(0)@BFvD2014"] =  0.31;
-            p["Lambda_b->Lambda::b_1_perp^A@BFvD2014"]  = -0.24;
-            p["mass::Lambda_b"]                         = 5.6194;
-            p["mass::Lambda"]                           = 1.1157;
-
-            TEST_CHECK_NEARLY_EQUAL(0.330, ff->f_long_v( 0.0), eps);
-            TEST_CHECK_NEARLY_EQUAL(0.418, ff->f_long_v( 5.0), eps);
-            TEST_CHECK_NEARLY_EQUAL(0.555, ff->f_long_v(10.0), eps);
-            TEST_CHECK_NEARLY_EQUAL(0.794, ff->f_long_v(15.0), eps);
-            TEST_CHECK_NEARLY_EQUAL(1.302, ff->f_long_v(20.0), eps);
-
-            TEST_CHECK_NEARLY_EQUAL(0.310, ff->f_long_a( 0.0), eps);
-            TEST_CHECK_NEARLY_EQUAL(0.369, ff->f_long_a( 5.0), eps);
-            TEST_CHECK_NEARLY_EQUAL(0.453, ff->f_long_a(10.0), eps);
-            TEST_CHECK_NEARLY_EQUAL(0.584, ff->f_long_a(15.0), eps);
-            TEST_CHECK_NEARLY_EQUAL(0.810, ff->f_long_a(20.0), eps);
-
-            TEST_CHECK_NEARLY_EQUAL(0.340, ff->f_perp_v( 0.0), eps);
-            TEST_CHECK_NEARLY_EQUAL(0.429, ff->f_perp_v( 5.0), eps);
-            TEST_CHECK_NEARLY_EQUAL(0.567, ff->f_perp_v(10.0), eps);
-            TEST_CHECK_NEARLY_EQUAL(0.806, ff->f_perp_v(15.0), eps);
-            TEST_CHECK_NEARLY_EQUAL(1.315, ff->f_perp_v(20.0), eps);
-
-            TEST_CHECK_NEARLY_EQUAL(0.310, ff->f_perp_a( 0.0), eps);
-            TEST_CHECK_NEARLY_EQUAL(0.366, ff->f_perp_a( 5.0), eps);
-            TEST_CHECK_NEARLY_EQUAL(0.446, ff->f_perp_a(10.0), eps);
-            TEST_CHECK_NEARLY_EQUAL(0.568, ff->f_perp_a(15.0), eps);
-            TEST_CHECK_NEARLY_EQUAL(0.780, ff->f_perp_a(20.0), eps);
-        }
-} bfvd2014_form_factors_test;
-
 class DM2016FormFactorsTest :
     public TestCase
 {
